+++ src/index.ts
@@ -1,3 +1,5 @@
+import { login } from './commands/login.js';
+import { logout } from './commands/logout.js';
 import { start } from './commands/start.js';
 import { version } from './commands/version.js';
 
@@ -5,7 +7,13 @@ const command = process.argv[2];
 
 switch (command) {
   case 'start':
-    start();
+    await start();
+    break;
+  case 'login':
+    await login();
+    break;
+  case 'logout':
+    await logout();
     break;
   case 'version':
   case '--version':
@@ -17,6 +25,8 @@ switch (command) {
     console.log('');
     console.log('Commands:');
     console.log('  start      Start the tmonier daemon');
+    console.log('  login      Save your API key');
+    console.log('  logout     Clear saved credentials');
     console.log('  version    Show version');
     break;
 }